+++ mvc/src/main/java/com/interface21/webmvc/servlet/mapping/HandlerKey.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.handler.mapping;
+package com.interface21.webmvc.servlet.mapping;
 
 import com.interface21.web.bind.annotation.RequestMethod;
 import java.util.Objects;
