+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.interface21.webmvc.servlet.mvc.handler.mapping.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.handler.mapping.HandlerExecution;
+import com.interface21.webmvc.servlet.mapping.AnnotationHandlerMapping;
+import com.interface21.webmvc.servlet.mapping.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
